fix: moved monitoringservice to its own folder
